+++ R/utils_link_inverse.R
@@ -1,15 +1,20 @@
 #' @importFrom stats family
 get_link_fun <- function(model) {
-  # get model family
-  ff <- stats::family(model)
+  tryCatch(
+    {
+      # get model family
+      ff <- stats::family(model)
 
-  # return link function, if exists
-  if ("linkfun" %in% names(ff)) return(ff$linkfun)
+      # return link function, if exists
+      if ("linkfun" %in% names(ff)) return(ff$linkfun)
 
-  # else, create link function from link-string
-  if ("link" %in% names(ff)) return(match.fun(ff$link))
-
-  NULL
+      # else, create link function from link-string
+      if ("link" %in% names(ff)) return(match.fun(ff$link))
+    },
+    error = function(x) { NULL },
+    warning = function(x) { NULL },
+    finally = function(x) { NULL }
+  )
 }